# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VirgoFin App Docs",
        "description": "API documentation for the Virgo Financial Application",
        "contact": {
            "name": "VirgoFin App Support",
            "email": "support@virgofinapp.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9030",
            "description": "The VirgoFin App API"
        },
        {
            "url": "http://localhost:9030/"
        }
    ],
    "paths": {
...
        "/api/orders/sides": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get order side options",
                "operationId": "ed83447bcfcae9a6dc7650a58a296d13",
                "responses": {
                    "200": {
                        "description": "Order side options retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Buy"
                                                    },
                                                    "value": {
                                                        "type": "string",
                                                        "example": "buy"
                                                    },
                                                    "selected": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/symbols": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get tradable symbol options",
                "operationId": "0f7f101bcd182670f3244178e146db59",
                "responses": {
                    "200": {
                        "description": "Symbol options retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "BTC"
                                                    },
                                                    "value": {
                                                        "type": "string",
                                                        "example": "BTC"
                                                    },
                                                    "selected": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use Laravel Sanctum for API authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}